+++ duckscript_sdk/src/sdk/std/collections/mod.rs
@@ -28,6 +28,12 @@ mod map_to_properties;
 mod range;
 mod read_properties;
 mod set;
+mod set_clear;
+mod set_contains;
+mod set_is_empty;
+mod set_put;
+mod set_remove;
+mod set_size;
 mod write_properties;
 
 use crate::utils::pckg;
@@ -69,6 +75,12 @@ pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptEr
     commands.set(range::create(&package))?;
     commands.set(read_properties::create(&package))?;
     commands.set(set::create(&package))?;
+    commands.set(set_clear::create(&package))?;
+    commands.set(set_contains::create(&package))?;
+    commands.set(set_is_empty::create(&package)?)?;
+    commands.set(set_put::create(&package))?;
+    commands.set(set_remove::create(&package))?;
+    commands.set(set_size::create(&package))?;
     commands.set(write_properties::create(&package))?;
 
     Ok(())